+++ lib/bibliotheca/collection.rb
@@ -5,8 +5,8 @@ end
 require_relative './collection/exercise_not_found_error'
 require_relative './collection/with_slug'
 require_relative './collection/guide_array'
-require_relative './collection/guide_collection'
 require_relative './collection/language_array'
 require_relative './collection/languages'
-require_relative './collection/book_collection'
+require_relative './collection/guides'
+require_relative './collection/books'
 require_relative './collection/book_array'
