+++ lib/bibliotheca/collection/books.rb
