+++ lib/bibliotheca/collection/guides.rb
